+++ src/main/java/com/example/demo/repository/UserRepository.java
@@ -9,7 +9,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-import com.example.demo.dto.UserDailyTimestampDTO;
 import com.example.demo.entity.User;
 
 public interface UserRepository extends JpaRepository<User, Long> {
@@ -19,7 +18,7 @@ public interface UserRepository extends JpaRepository<User, Long> {
 
 	void deleteById(Long userId);
 	User findByEmail(String email);
-	User findByEmployeeNo(Integer employeeNo);
+	User findByEmployeeNo(Long employeeNo);
 
 	@Query(value = """
             SELECT
